==> analysis/1-biocontrol_history.R <==
require(here)
require(sf)
require(dplyr)
require(sp)
require(lubridate)
setwd(paste0(here(), '/../data/'))
#private parasitoid release data - see mapbiocontrol.org for permissions
release_sites<-read_sf('./Extract_Data_Jun_11__2021__10_41_46_AM/Parasitoid_Release.shp')
recovery_sites<-read_sf('./Extract_Data_Jun_11__2021__10_41_46_AM/Recovery.shp')
grid<-read.csv('countydatanorm_march.csv')
coordinates(grid)<-~X_coord+Y_coord
grid<-st_as_sf(grid)
st_crs(grid)<-'+proj=eqdc +lat_0=40 +lon_0=-96 +lat_1=20 +lat_2=60 +x_0=0 +y_0=0 +datum=NAD83 +units=m +no_defs'
release_sites<-st_transform(release_sites, '+proj=eqdc +lat_0=40 +lon_0=-96 +lat_1=20 +lat_2=60 +x_0=0 +y_0=0 +datum=NAD83 +units=m +no_defs')
grid<-grid%>%select(ID)
release_sites<-release_sites%>%st_join(grid, join=st_nearest_feature)
recovery_sites<-st_transform(recovery_sites, '+proj=eqdc +lat_0=40 +lon_0=-96 +lat_1=20 +lat_2=60 +x_0=0 +y_0=0 +datum=NAD83 +units=m +no_defs')
recovery_sites<-recovery_sites%>%st_join(grid, join=st_nearest_feature)
release_sites$year<-as.numeric(gsub('-.*','',release_sites$RELEASE_DA))
release_sites$fiveyear<-ceiling_date(as.Date(ISOdate(release_sites$year, 1,1)), '5year')
release_sum<-release_sites%>%group_by(ID, fiveyear)%>%summarize_if(is.numeric, sum)

recovery_sites$year<-as.numeric(gsub('-.*','',recovery_sites$DATE_SAMPL))
recovery_sites$fiveyear<-ceiling_date(as.Date(ISOdate(recovery_sites$year, 1,1)), '5year')
recovery_sum<-recovery_sites%>%group_by(ID, fiveyear)%>%summarize_if(is.numeric, sum)
recovery_sum<-subset(recovery_sum,fiveyear!='2025-01-01')
release_sum<-subset(release_sum,fiveyear!='2025-01-01')
release_sum$total<-rowSums(st_drop_geometry(release_sum[,grep('COUNT',colnames(release_sum))]))
release_sum<-subset(release_sum, total>10000)
recovery_sum<-subset(recovery_sum, ID%in%release_sum$ID)
recovery_sum$total<-rowSums(st_drop_geometry(recovery_sum[,grep('NUM',colnames(recovery_sum))]))
recovery_sum<-subset(recovery_sum, total>100)
release_sum$ID_year<-paste0(release_sum$ID, release_sum$fiveyear)
recovery_sum$ID_year<-paste0(recovery_sum$ID, recovery_sum$fiveyear)
release_sum<-subset(release_sum, ID_year%in%recovery_sum$ID_year)

release_sum$fiveyear<-as.numeric(gsub('-.*','',release_sum$fiveyear))
write.csv(cbind(release_sum$ID, release_sum$fiveyear, match(release_sum$ID,data$ID)), file="biocontrol_history.csv", row.names=F)
